backend/requirements.txt
# Base requirements
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
django-filter>=23.0.0
Pillow>=9.5.0
python-decouple>=3.8.0

# Database
psycopg2-binary>=2.9.0

# Development
django-extensions>=3.2.0

# AI/ML
google-generativeai>=0.3.0

# Production
gunicorn>=20.1.0
whitenoise>=6.4.0